Update Update-ModulesInAutomationToLatestVersion.ps1
Added an if condition for checking the current version of AzureRM.Resources Module. This is required in cases where an automation account is created without specifying any specific module versions, then default module version of 1.0.3 is imported. In this case, if this runbook is triggered with Get-AzureRmResource cmdlet, then it fails with an error "Parameter set cannot be resolved using the specified named parameters.". Adding this check fixes this issue and updates the modules